Simplify caching policy resolution with concrete object contraint
