--- src/CQRS/LeanCode.CQRS.OutputCaching/CacheOutputAttribute.cs
@@ -1,5 +1,0 @@
-namespace LeanCode.CQRS.OutputCaching;
-
-// TODO: Port this to `LeanCode.Contracts` and use this attribute everywhere
-[AttributeUsage(AttributeTargets.Class, Inherited = false)]
-public sealed class CacheOutputAttribute : Attribute;
